# Reduce allocations in metrics and tracing