Switch to call `metric.WithAttributeSet` instead of `metric.WithAttributes` to reduce allocations. Set trace option slice capacity to avoid slice resizing. Avoid using `strings.Fields` in favor of `TrimSpace`/`IndexFunc`.