@pkwarren 2025-04-24
Results w/ newly added benchmark:
```
goos: darwin
goarch: arm64
pkg: github.com/exaring/otelpgx
cpu: Apple M1 Max
          │   old.txt   │            new.txt            │
          │   sec/op    │   sec/op     vs base          │
Tracer-10   1.241m ± 9%   1.168m ± 6%  ~ (p=0.315 n=10)

          │    old.txt    │               new.txt               │
          │     B/op      │     B/op      vs base               │
Tracer-10   10.125Ki ± 0%   9.641Ki ± 0%  -4.79% (p=0.000 n=10)

          │  old.txt   │              new.txt              │
          │ allocs/op  │ allocs/op   vs base               │
Tracer-10   96.00 ± 0%   89.00 ± 0%  -7.29% (p=0.000 n=10)
```