@pkwarren 2025-05-02
@obitech @costela - Sorry to ping you directly, but are either of you a maintainer who could review this PR? We're happy users of otelpgx but saw a few opportunities to reduce allocations.